#1 (3)

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,7 +1,4 @@
-#include "function_pointers.h"
-#include <stdio.h>
 #include <stddef.h>
-#include <stdlib.h>
 /**
  * array_iterator - function that prints name
  * @array: array
@@ -11,7 +8,7 @@
  **/
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if ((array == NULL) || (size == 0) || (action == NULL))
 		return;
